requirements.txt
openpyxl>=3.1.2
openai>=1.30.0
python-dotenv>=1.0.1
